move /generated to /contracts
