--- cmd/ebrelayer/contract/abi/CosmosBridge.abi
@@ -1,1 +1,0 @@
-[{"inputs":[{"internalType":"address","name":"_operator","type":"address"},{"internalType":"address","name":"_valset","type":"address"}],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"_bridgeBank","type":"address"}],"name":"LogBridgeBankSet","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_prophecyID","type":"uint256"},{"indexed":false,"internalType":"enum CosmosBridge.ClaimType","name":"_claimType","type":"uint8"},{"indexed":false,"internalType":"bytes","name":"_cosmosSender","type":"bytes"},{"indexed":false,"internalType":"address payable","name":"_ethereumReceiver","type":"address"},{"indexed":false,"internalType":"address","name":"_validatorAddress","type":"address"},{"indexed":false,"internalType":"address","name":"_tokenAddress","type":"address"},{"indexed":false,"internalType":"string","name":"_symbol","type":"string"},{"indexed":false,"internalType":"uint256","name":"_amount","type":"uint256"}],"name":"LogNewProphecyClaim","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"_oracle","type":"address"}],"name":"LogOracleSet","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_prophecyID","type":"uint256"},{"indexed":false,"internalType":"enum CosmosBridge.ClaimType","name":"_claimType","type":"uint8"}],"name":"LogProphecyCompleted","type":"event"},{"constant":true,"inputs":[],"name":"bridgeBank","outputs":[{"internalType":"contract BridgeBank","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"hasBridgeBank","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"hasOracle","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"operator","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"oracle","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"prophecyClaimCount","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"prophecyClaims","outputs":[{"internalType":"enum CosmosBridge.ClaimType","name":"claimType","type":"uint8"},{"internalType":"bytes","name":"cosmosSender","type":"bytes"},{"internalType":"address payable","name":"ethereumReceiver","type":"address"},{"internalType":"address","name":"originalValidator","type":"address"},{"internalType":"address","name":"tokenAddress","type":"address"},{"internalType":"string","name":"symbol","type":"string"},{"internalType":"uint256","name":"amount","type":"uint256"},{"internalType":"enum CosmosBridge.Status","name":"status","type":"uint8"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"valset","outputs":[{"internalType":"contract Valset","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"_oracle","type":"address"}],"name":"setOracle","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address payable","name":"_bridgeBank","type":"address"}],"name":"setBridgeBank","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"enum CosmosBridge.ClaimType","name":"_claimType","type":"uint8"},{"internalType":"bytes","name":"_cosmosSender","type":"bytes"},{"internalType":"address payable","name":"_ethereumReceiver","type":"address"},{"internalType":"address","name":"_tokenAddress","type":"address"},{"internalType":"string","name":"_symbol","type":"string"},{"internalType":"uint256","name":"_amount","type":"uint256"}],"name":"newProphecyClaim","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"}],"name":"completeProphecyClaim","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"}],"name":"isProphecyClaimActive","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"}],"name":"isProphecyClaimValidatorActive","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"}]
