--- cmd/ebrelayer/contract/abi/Oracle.abi
@@ -1,1 +1,0 @@
-[{"inputs":[{"internalType":"address","name":"_operator","type":"address"},{"internalType":"address","name":"_valset","type":"address"},{"internalType":"address","name":"_cosmosBridge","type":"address"}],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_prophecyID","type":"uint256"},{"indexed":false,"internalType":"string","name":"_message","type":"string"},{"indexed":false,"internalType":"address","name":"_validatorAddress","type":"address"},{"indexed":false,"internalType":"bytes","name":"_signature","type":"bytes"}],"name":"LogNewOracleClaim","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_prophecyID","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_weightedSignedPower","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_weightedTotalPower","type":"uint256"},{"indexed":false,"internalType":"address","name":"_submitter","type":"address"}],"name":"LogProphecyProcessed","type":"event"},{"constant":true,"inputs":[],"name":"cosmosBridge","outputs":[{"internalType":"contract CosmosBridge","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"address","name":"","type":"address"}],"name":"hasMadeClaim","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"operator","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"uint256","name":"","type":"uint256"}],"name":"oracleClaimValidators","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"valset","outputs":[{"internalType":"contract Valset","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"},{"internalType":"string","name":"_message","type":"string"},{"internalType":"bytes","name":"_signature","type":"bytes"}],"name":"newOracleClaim","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"}],"name":"processBridgeProphecy","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"internalType":"uint256","name":"_prophecyID","type":"uint256"}],"name":"checkBridgeProphecy","outputs":[{"internalType":"bool","name":"","type":"bool"},{"internalType":"uint256","name":"","type":"uint256"},{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"}]
