--- cmd/ebrelayer/contract/abi/Valset.abi
@@ -1,1 +1,0 @@
-[{"inputs":[{"internalType":"address","name":"_operator","type":"address"},{"internalType":"address[]","name":"_initValidators","type":"address[]"},{"internalType":"uint256[]","name":"_initPowers","type":"uint256[]"}],"payable":false,"stateMutability":"nonpayable","type":"constructor"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"_validator","type":"address"},{"indexed":false,"internalType":"uint256","name":"_power","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_currentValsetVersion","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_validatorCount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_totalPower","type":"uint256"}],"name":"LogValidatorAdded","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"_validator","type":"address"},{"indexed":false,"internalType":"uint256","name":"_power","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_currentValsetVersion","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_validatorCount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_totalPower","type":"uint256"}],"name":"LogValidatorPowerUpdated","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"address","name":"_validator","type":"address"},{"indexed":false,"internalType":"uint256","name":"_power","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_currentValsetVersion","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_validatorCount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_totalPower","type":"uint256"}],"name":"LogValidatorRemoved","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_newValsetVersion","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_validatorCount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_totalPower","type":"uint256"}],"name":"LogValsetReset","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"uint256","name":"_newValsetVersion","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_validatorCount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"_totalPower","type":"uint256"}],"name":"LogValsetUpdated","type":"event"},{"constant":true,"inputs":[],"name":"currentValsetVersion","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"operator","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"powers","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"totalPower","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[],"name":"validatorCount","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"name":"validators","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"string","name":"_message","type":"string"},{"internalType":"bytes","name":"_signature","type":"bytes"}],"name":"recover","outputs":[{"internalType":"address","name":"","type":"address"}],"payable":false,"stateMutability":"pure","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"_validatorAddress","type":"address"},{"internalType":"uint256","name":"_validatorPower","type":"uint256"}],"name":"addValidator","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"_validatorAddress","type":"address"},{"internalType":"uint256","name":"_newValidatorPower","type":"uint256"}],"name":"updateValidatorPower","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address","name":"_validatorAddress","type":"address"}],"name":"removeValidator","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":false,"inputs":[{"internalType":"address[]","name":"_validators","type":"address[]"},{"internalType":"uint256[]","name":"_powers","type":"uint256[]"}],"name":"updateValset","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"},{"constant":true,"inputs":[{"internalType":"address","name":"_validatorAddress","type":"address"}],"name":"isActiveValidator","outputs":[{"internalType":"bool","name":"","type":"bool"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":true,"inputs":[{"internalType":"address","name":"_validatorAddress","type":"address"}],"name":"getValidatorPower","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"payable":false,"stateMutability":"view","type":"function"},{"constant":false,"inputs":[{"internalType":"uint256","name":"_valsetVersion","type":"uint256"},{"internalType":"address","name":"_validatorAddress","type":"address"}],"name":"recoverGas","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"}]
