+++ cmd/ebrelayer/contract/generator.go
@@ -16,17 +16,17 @@ var (
 	// BaseABIBINGenCmd is the base command for contract compilation to ABI and BIN
 	BaseABIBINGenCmd = strings.Join([]string{"solc ",
 		fmt.Sprintf("--%s ./testnet-contracts/contracts/%s%s.sol ", SolcCmdText, DirectoryText, ContractText),
-		fmt.Sprintf("-o ./cmd/ebrelayer/generated/%s/%s ", SolcCmdText, ContractText),
+		fmt.Sprintf("-o ./cmd/ebrelayer/contract/generated/%s/%s ", SolcCmdText, ContractText),
 		"--overwrite ",
 		"--allow-paths *,"},
 		"")
 	// BaseBindingGenCmd is the base command for contract binding generation
 	BaseBindingGenCmd = strings.Join([]string{"abigen ",
-		fmt.Sprintf("--bin ./cmd/ebrelayer/generated/bin/%s/%s.bin ", ContractText, ContractText),
-		fmt.Sprintf("--abi ./cmd/ebrelayer/generated/abi/%s/%s.abi ", ContractText, ContractText),
+		fmt.Sprintf("--bin ./cmd/ebrelayer/contract/generated/bin/%s/%s.bin ", ContractText, ContractText),
+		fmt.Sprintf("--abi ./cmd/ebrelayer/contract/generated/abi/%s/%s.abi ", ContractText, ContractText),
 		fmt.Sprintf("--pkg %s ", ContractText),
 		fmt.Sprintf("--type %s ", ContractText),
-		fmt.Sprintf("--out ./cmd/ebrelayer/generated/bindings/%s/%s.go", ContractText, ContractText)},
+		fmt.Sprintf("--out ./cmd/ebrelayer/contract/generated/bindings/%s/%s.go", ContractText, ContractText)},
 		"")
 )
